feat: extract shared library for backup services
