+++ ResguardoApp/DiscoRespaldoInfo.cs
@@ -1,3 +1,4 @@
+using SharedLib;
 using System;
 using System.Runtime.InteropServices;
 using System.Management;
