+++ ResguardoApp/MainForm.cs
@@ -1,3 +1,4 @@
+using SharedLib;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
