+++ ResguardoApp/ResguardoApp.csproj
@@ -15,5 +15,8 @@
     <PackageReference Include="System.ServiceProcess.ServiceController" Version="6.0.0" />
     <PackageReference Include="System.Text.Json" Version="9.0.8" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../SharedLib/SharedLib.csproj" />
+  </ItemGroup>
 
 </Project>
