+++ ResguardoAppService/Service1.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.ServiceProcess;
 using System.Diagnostics;
 using System.Timers;
-using ResguardoApp;
+using SharedLib;
 //using System.Text.Json;
 using Newtonsoft.Json;
