+++ SharedLib/BackupService.cs
@@ -1,9 +1,8 @@
 using System;
 using System.IO;
 using System.Linq;
-using System.Windows.Forms;
 
-namespace ResguardoApp
+namespace SharedLib
 {
     public static class BackupService
     {
